WIP geolocation improvements

+++ web/src/types/osdm.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize)]
 pub struct OsdmGeoPosition {
-    pub latitude: f64,
-    pub longitude: f64,
+    pub latitude: f32,
+    pub longitude: f32,
 }
 
 #[derive(Deserialize, Serialize)]
